+++ src/main/java/fr/maxlego08/essentials/module/modules/MailBoxModule.java
@@ -194,4 +194,13 @@ public class MailBoxModule extends ZModule {
 
         message(sender, Message.MAILBOX_GIVE_ALL, "%item%", itemName, "%amount%", amount);
     }
+
+    public void clear(CommandSender sender, UUID uuid, String username) {
+        getStorage().clearMailBox(uuid);
+        var user = getUser(uuid);
+        if (user != null) {
+            user.setMailBoxItems(List.of());
+        }
+        message(sender, Message.MAILBOX_CLEAR, "%player%", username);
+    }
 }